requirements.txt
streamlit
python-dotenv
requests
google-generativeai
groq
